requirements.txt
streamlit
openai
elevenlabs
firecrawl-py
python-dotenv